Remove rmgpy.reaction.ReactionModel class and merge functionality into rmg.model.ReactionModel
Retain only the merge and reduce functions.  Eliminate the
generateStoichiometryMatrix function, which does not appear
to be used anywhere.
